# Make bytesPipe use linear allocations