--- a/pkg/ioutils/bytespipe.go
+++ b/pkg/ioutils/bytespipe.go
@@ -1,6 +1,6 @@
 package ioutils
 
-const maxCap = 10 * 1e6
+const maxCap = 1e6
 
 // BytesPipe is io.ReadWriter which works similary to pipe(queue).
 // All written data could be read only once. Also BytesPipe trying to adjust
@@ -8,8 +8,11 @@ const maxCap = 10 * 1e6
 // after highload peak.
 // BytesPipe isn't goroutine-safe, caller must synchronize it if needed.
 type BytesPipe struct {
-	buf      []byte
-	lastRead int
+	buf        [][]byte // slice of byte-slices of buffered data
+	lastRead   int      // index in the first slice to a read point
+	bufLen     int      // length of data buffered over the slices
+	bufCap     int      // whole allocated capacity over slices
+	writeIndex int      // index to the buffer where the writes go
 }
 
 // NewBytesPipe creates new BytesPipe, initialized by specified slice.
@@ -20,26 +23,20 @@ func NewBytesPipe(buf []byte) *BytesPipe {
 		buf = make([]byte, 0, 64)
 	}
 	return &BytesPipe{
-		buf: buf[:0],
+		buf:    [][]byte{buf[:0]},
+		bufCap: cap(buf),
 	}
 }
 
 func (bp *BytesPipe) grow(n int) {
-	if len(bp.buf)+n > cap(bp.buf) {
+	for bp.bufLen+n > bp.bufCap {
 		// not enough space
-		var buf []byte
-		remain := bp.len()
-		if remain+n <= cap(bp.buf)/2 {
-			// enough space in current buffer, just move data to head
-			copy(bp.buf, bp.buf[bp.lastRead:])
-			buf = bp.buf[:remain]
-		} else {
-			// reallocate buffer
-			buf = make([]byte, remain, 2*cap(bp.buf)+n)
-			copy(buf, bp.buf[bp.lastRead:])
+		nextCap := 2 * cap(bp.buf[len(bp.buf)-1])
+		if maxCap < nextCap {
+			nextCap = maxCap
 		}
-		bp.buf = buf
-		bp.lastRead = 0
+		bp.buf = append(bp.buf, make([]byte, 0, nextCap))
+		bp.bufCap += nextCap
 	}
 }
 
@@ -47,36 +44,50 @@ func (bp *BytesPipe) grow(n int) {
 // It can increase cap of internal []byte slice in a process of writing.
 func (bp *BytesPipe) Write(p []byte) (n int, err error) {
 	bp.grow(len(p))
-	bp.buf = append(bp.buf, p...)
+	for {
+		b := bp.buf[bp.writeIndex]
+		n := copy(b[len(b):cap(b)], p)
+		bp.bufLen += n
+		bp.buf[bp.writeIndex] = b[:len(b)+n]
+		if len(p) > n { // more data: write to the next slice
+			p = p[n:]
+			bp.writeIndex += 1
+			continue
+		}
+		break
+	}
 	return
 }
 
 func (bp *BytesPipe) len() int {
-	return len(bp.buf) - bp.lastRead
-}
-
-func (bp *BytesPipe) crop() {
-	// shortcut for empty buffer
-	if bp.lastRead == len(bp.buf) {
-		bp.lastRead = 0
-		bp.buf = bp.buf[:0]
-	}
-	r := bp.len()
-	// if we have too large buffer for too small data
-	if cap(bp.buf) > maxCap && r < cap(bp.buf)/10 {
-		copy(bp.buf, bp.buf[bp.lastRead:])
-		// will use same underlying slice until reach cap
-		bp.buf = bp.buf[:r : cap(bp.buf)/2]
-		bp.lastRead = 0
-	}
+	return bp.bufLen - bp.lastRead
 }
 
 // Read reads bytes from BytesPipe.
 // Data could be read only once.
-// Internal []byte slice could be shrinked.
 func (bp *BytesPipe) Read(p []byte) (n int, err error) {
-	n = copy(p, bp.buf[bp.lastRead:])
-	bp.lastRead += n
-	bp.crop()
+	for {
+		written := copy(p, bp.buf[0][bp.lastRead:])
+		n += written
+		bp.lastRead += written
+		if len(p) > written && bp.len() > 0 {
+			// more buffered data and more asked. read from next slice.
+			p = p[written:]
+			bp.lastRead = 0
+			bp.bufLen -= len(bp.buf[0])
+			bp.bufCap -= len(bp.buf[0])
+			bp.buf[0] = nil     // throw away old slice
+			bp.buf = bp.buf[1:] // switch to next
+			bp.writeIndex -= 1
+			continue
+		}
+		if bp.len() == 0 {
+			// we have read everything. reset to the beginning.
+			bp.lastRead = 0
+			bp.bufLen -= len(bp.buf[0])
+			bp.buf[0] = bp.buf[0][:0]
+		}
+		break
+	}
 	return
 }