@LK4D4 2015-09-17
@tonistiigi I wonder why in first version there was no such difference in allocs. Is it because maxCap?