@tonistiigi 2015-09-17
@LK4D4 Yes. In first version grow always took double of previous cap or even more if this wasn't enough. If we put it back then allocs are same as in your code, bytes is 10-20% more than this version.